Se agrega un mensaje en un modal cuando los campos no se pueden eliminar debido a que son claves foráneas de otra tabla. Dentro del modal se agrega un botón que los lleva al registro mensionado en el error.

+++ site-php/includes/Comisarias.class.php
@@ -49,17 +49,30 @@
             $database = new Database();
             $conn = $database->getConnection();
 
-            $stmt = $conn->prepare('DELETE FROM cctv_comisarias WHERE id=:id');
-            $stmt->bindParam(':id',$id);
-            if($stmt->execute()){
-                return [
-                    'status' => true,
-                    'message' => 'comisaria eliminada correctamente'
-                ];
+            $stmt = $conn->prepare('SELECT * FROM cctv_plantas WHERE id_comisarias = :id');
+            $stmt->bindParam(':id', $id);
+            $stmt->execute();
+            $plantasTipoPlanta = $stmt->fetchAll(PDO::FETCH_ASSOC);
+
+            if (empty($plantasTipoPlanta)) {
+
+                $stmt = $conn->prepare('DELETE FROM cctv_comisarias WHERE id=:id');
+                $stmt->bindParam(':id',$id);
+                if($stmt->execute()){
+                    return [
+                        'status' => true,
+                        'message' => 'comisaria eliminada correctamente'
+                    ];
+                } else {
+                    return [
+                        'status' => false,
+                        'message' => 'error al eliminar comisaria correctamente'
+                    ];
+                }
             } else {
                 return [
                     'status' => false,
-                    'message' => 'error al eliminar comisaria correctamente'
+                    'plantas' => $plantasTipoPlanta
                 ];
             }
         }
